requirements.txt
azure-core==1.13.0
azure-storage-blob==12.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.4.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
